fix(uht): wrap EVehicleType as EVehicleTypeBP for UPROPERTY use in VehicleSlotActor
UHT error from UE5.7 build:
  VehicleSlotActor.h(29): Error: Unable to find 'class', 'delegate', 'enum',
  or 'struct' with name 'EVehicleType'

Root cause: EVehicleType lives in TaraSimCore (Core-only, no UENUM macro
available — the invariant). UHT requires UENUM(BlueprintType) types for any
UPROPERTY field exposed to the reflection system.

Same problem already solved once: ESeason → EShedActionSeason wrapper in
ShedActionMenu.h. Applying the same pattern here:

EVehicleTypeBP : uint8 (UENUM, BlueprintType) — mirrors EVehicleType 1:1.
Call sites cast via static_cast<EVehicleType>((uint8)slotType) when invoking
bridge actions. Designer gets a proper dropdown in the Editor + UHT is happy.

Scanned the rest of TaraGame headers for other UPROPERTY references to
TaraSimCore enums — none. Only VehicleSlotActor had this issue.

+++ Source/TaraGame/Private/Actors/VehicleSlotActor.cpp
@@ -27,7 +27,7 @@ bool AVehicleSlotActor::IsInteractAvailable_Implementation() const
 	{
 		if (UTaraSimSubsystem* Sub = GI->GetSubsystem<UTaraSimSubsystem>())
 		{
-			return Sub->IsVehicleOwned((int32)SlotVehicleType);
+			return Sub->IsVehicleOwned((int32)(uint8)SlotVehicleType);
 		}
 	}
 	return false;
@@ -38,7 +38,7 @@ bool AVehicleSlotActor::OnInteract_Implementation(APawn* /*InstigatorPawn*/)
 	// Default: log + return true. Blueprint override does the possession
 	// swap to the actual vehicle Pawn (BP_QuadBike, BP_Bike, etc.).
 	UE_LOG(LogTemp, Display, TEXT("[VehicleSlot] Player chose vehicle type %d (override OnInteract in BP to handle)"),
-		(int32)SlotVehicleType);
+		(int32)(uint8)SlotVehicleType);
 	return true;
 }
